+++ lib/smart_id/errors/document_unusable_error.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module SmartId
+  module Errors
+    class DocumentUnusableError < Error
+      def initialize(message = "Document is unusable")
+        super
+      end
+    end
+  end
+end